+++ include/rrenvironment/microprocessor.h
@@ -20,18 +20,4 @@
 #define RR_ADDR_I2C_MC2 0x09 /* Ultrasonic and GPS */
 #define RR_ADDR_I2C_MS3 0x68 /* Gyroscope and accelometer */
 
-// Structure to send data to environment, used by Fat Controller
-class rr_io_tx {
-public:
-    uint8_t addr; // I2C address as defined above.
-    uint8_t io;   // this must be a value a RR_IO
-    std::list<uint8_t> bytes; // bytes to send the microprocessor
-};
-
-class rr_io_rx {
-    uint8_t io;   // this must be a value a RR_IO
-    std::list<uint8_t> bytes; // bytes recieved from micro processor.
-};
-
-
 #endif
